+++ dotnet/src/webdriver/Chrome/ChromeDriver.cs
@@ -18,6 +18,7 @@
 
 using System;
 using OpenQA.Selenium.Remote;
+using OpenQA.Selenium.Chromium;
 using System.Collections.Generic;
 
 namespace OpenQA.Selenium.Chrome
@@ -55,24 +56,14 @@ namespace OpenQA.Selenium.Chrome
     /// }
     /// </code>
     /// </example>
-    public class ChromeDriver : RemoteWebDriver
+    public class ChromeDriver : ChromiumDriver
     {
+
         /// <summary>
-        /// Accept untrusted SSL Certificates
+        /// Initializes a new instance of the <see cref="ChromeDriver"/> class.
         /// </summary>
-        public static readonly bool AcceptUntrustedCertificates = true;
-
-        private const string GetNetworkConditionsCommand = "getNetworkConditions";
-        private const string SetNetworkConditionsCommand = "setNetworkConditions";
-        private const string DeleteNetworkConditionsCommand = "deleteNetworkConditions";
-        private const string SendChromeCommand = "sendChromeCommand";
-        private const string SendChromeCommandWithResult = "sendChromeCommandWithResult";
-
-    /// <summary>
-    /// Initializes a new instance of the <see cref="ChromeDriver"/> class.
-    /// </summary>
-    public ChromeDriver()
-            : this(new ChromeOptions())
+        public ChromeDriver()
+                : this(new ChromeOptions())
         {
         }
 
@@ -145,108 +136,9 @@ namespace OpenQA.Selenium.Chrome
         /// <param name="options">The <see cref="ChromeOptions"/> to be used with the Chrome driver.</param>
         /// <param name="commandTimeout">The maximum amount of time to wait for each command.</param>
         public ChromeDriver(ChromeDriverService service, ChromeOptions options, TimeSpan commandTimeout)
-            : base(new DriverServiceCommandExecutor(service, commandTimeout), ConvertOptionsToCapabilities(options))
+            : base(service, options, commandTimeout)
         {
-            // Add the custom commands unique to Chrome
-            this.AddCustomChromeCommand(GetNetworkConditionsCommand, CommandInfo.GetCommand, "/session/{sessionId}/chromium/network_conditions");
-            this.AddCustomChromeCommand(SetNetworkConditionsCommand, CommandInfo.PostCommand, "/session/{sessionId}/chromium/network_conditions");
-            this.AddCustomChromeCommand(DeleteNetworkConditionsCommand, CommandInfo.DeleteCommand, "/session/{sessionId}/chromium/network_conditions");
-            this.AddCustomChromeCommand(SendChromeCommand, CommandInfo.PostCommand, "/session/{sessionId}/chromium/send_command");
-            this.AddCustomChromeCommand(SendChromeCommandWithResult, CommandInfo.PostCommand, "/session/{sessionId}/chromium/send_command_and_get_result");
         }
 
-        /// <summary>
-        /// Gets or sets the <see cref="IFileDetector"/> responsible for detecting
-        /// sequences of keystrokes representing file paths and names.
-        /// </summary>
-        /// <remarks>The Chrome driver does not allow a file detector to be set,
-        /// as the server component of the Chrome driver (ChromeDriver.exe) only
-        /// allows uploads from the local computer environment. Attempting to set
-        /// this property has no effect, but does not throw an exception. If you
-        /// are attempting to run the Chrome driver remotely, use <see cref="RemoteWebDriver"/>
-        /// in conjunction with a standalone WebDriver server.</remarks>
-        public override IFileDetector FileDetector
-        {
-            get { return base.FileDetector; }
-            set { }
-        }
-
-        /// <summary>
-        /// Gets or sets the network condition emulation for Chrome.
-        /// </summary>
-        public ChromeNetworkConditions NetworkConditions
-        {
-            get
-            {
-                Response response = this.Execute(GetNetworkConditionsCommand, null);
-                return ChromeNetworkConditions.FromDictionary(response.Value as Dictionary<string, object>);
-            }
-
-            set
-            {
-                if (value == null)
-                {
-                    throw new ArgumentNullException("value", "value must not be null");
-                }
-
-                Dictionary<string, object> parameters = new Dictionary<string, object>();
-                parameters["network_conditions"] = value.ToDictionary();
-                this.Execute(SetNetworkConditionsCommand, parameters);
-            }
-        }
-
-        /// <summary>
-        /// Executes a custom Chrome command.
-        /// </summary>
-        /// <param name="commandName">Name of the command to execute.</param>
-        /// <param name="commandParameters">Parameters of the command to execute.</param>
-        public void ExecuteChromeCommand(string commandName, Dictionary<string, object> commandParameters)
-        {
-            if (commandName == null)
-            {
-                throw new ArgumentNullException("commandName", "commandName must not be null");
-            }
-
-            Dictionary<string, object> parameters = new Dictionary<string, object>();
-            parameters["cmd"] = commandName;
-            parameters["params"] = commandParameters;
-            this.Execute(SendChromeCommand, parameters);
-        }
-
-        /// <summary>
-        /// Executes a custom Chrome command that returns a result.
-        /// </summary>
-        /// <param name="commandName">Name of the command to execute.</param>
-        /// <param name="commandParameters">Parameters of the command to execute.</param>
-        /// <returns>The JSON return value of the Chrome command, converted to a .NET object.</returns>
-        public object ExecuteChromeCommandWithResult(string commandName, Dictionary<string, object> commandParameters)
-        {
-            if (commandName == null)
-            {
-                throw new ArgumentNullException("commandName", "commandName must not be null");
-            }
-
-            Dictionary<string, object> parameters = new Dictionary<string, object>();
-            parameters["cmd"] = commandName;
-            parameters["params"] = commandParameters;
-            Response response = this.Execute(SendChromeCommandWithResult, parameters);
-            return response.Value;
-        }
-
-        private static ICapabilities ConvertOptionsToCapabilities(ChromeOptions options)
-        {
-            if (options == null)
-            {
-                throw new ArgumentNullException("options", "options must not be null");
-            }
-
-            return options.ToCapabilities();
-        }
-
-        private void AddCustomChromeCommand(string commandName, string method, string resourcePath)
-        {
-            CommandInfo commandInfoToAdd = new CommandInfo(method, resourcePath);
-            this.CommandExecutor.CommandInfoRepository.TryAddCommand(commandName, commandInfoToAdd);
-        }
     }
 }
