+++ dotnet/src/webdriver/Chromium/ChromiumWebElement.cs
@@ -16,21 +16,21 @@
 // limitations under the License.
 // </copyright>
 
-using OpenQA.Selenium.Chromium;
+using OpenQA.Selenium.Remote;
 
-namespace OpenQA.Selenium.Chrome
+namespace OpenQA.Selenium.Chromium
 {
     /// <summary>
     /// Provides a mechanism to get elements off the page for test
     /// </summary>
-    public class ChromeWebElement : ChromiumWebElement
+    public abstract class ChromiumWebElement : RemoteWebElement
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="ChromeWebElement"/> class.
+        /// Initializes a new instance of the <see cref="ChromiumWebElement"/> class.
         /// </summary>
         /// <param name="parent">Driver in use</param>
         /// <param name="elementId">Id of the element</param>
-        public ChromeWebElement(ChromeDriver parent, string elementId)
+        public ChromiumWebElement(ChromiumDriver parent, string elementId)
             : base(parent, elementId)
         {
         }
