+++ dotnet/src/webdriver/Edge/EdgeDriverService.cs
@@ -20,20 +20,23 @@ using System;
 using System.Globalization;
 using System.Text;
 using OpenQA.Selenium.Internal;
+using OpenQA.Selenium.Chromium;
 
 namespace OpenQA.Selenium.Edge
 {
     /// <summary>
-    /// Exposes the service provided by the native MicrosoftWebDriver executable.
+    /// Exposes the service provided by the native WebDriver executable.
     /// </summary>
-    public sealed class EdgeDriverService : DriverService
+    public sealed class EdgeDriverService : ChromiumDriverService
     {
         private const string MicrosoftWebDriverServiceFileName = "MicrosoftWebDriver.exe";
-        private static readonly Uri MicrosoftWebDriverDownloadUrl = new Uri("http://go.microsoft.com/fwlink/?LinkId=619687");
+        private static readonly Uri MicrosoftWebDriverDownloadUrl = new Uri("https://developer.microsoft.com/en-us/microsoft-edge/tools/webdriver/");
         private string host;
         private string package;
         private bool useVerboseLogging;
         private bool? useSpecCompliantProtocol;
+        private const string MSEdgeDriverServiceFileName = "MSEdgeDriver.exe";
+        private bool isLegacy;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="EdgeDriverService"/> class.
@@ -41,41 +44,95 @@ namespace OpenQA.Selenium.Edge
         /// <param name="executablePath">The full path to the EdgeDriver executable.</param>
         /// <param name="executableFileName">The file name of the EdgeDriver executable.</param>
         /// <param name="port">The port on which the EdgeDriver executable should listen.</param>
-        private EdgeDriverService(string executablePath, string executableFileName, int port)
-            : base(executablePath, port, executableFileName, MicrosoftWebDriverDownloadUrl)
+        /// <param name="isLegacy">Whether to use legacy mode to launch Edge.</param>
+        private EdgeDriverService(string executablePath, string executableFileName, int port, bool isLegacy)
+            : base(executablePath, executableFileName, port, MicrosoftWebDriverDownloadUrl)
         {
+            this.isLegacy = isLegacy;
         }
 
         /// <summary>
         /// Gets or sets the value of the host adapter on which the Edge driver service should listen for connections.
+        /// The property only exists in legacy mode.
         /// </summary>
         public string Host
         {
-            get { return this.host; }
-            set { this.host = value; }
+            get
+            {
+                if (!this.isLegacy)
+                {
+                    throw new ArgumentException("Host property does not exist");
+                }
+
+                return this.host;
+            }
+            set
+            {
+                if (!this.isLegacy)
+                {
+                    throw new ArgumentException("Host property does not exist");
+                }
+
+                this.host = value;
+            }
         }
 
         /// <summary>
         /// Gets or sets the value of the package the Edge driver service will launch and automate.
+        /// The property only exists in legacy mode.
         /// </summary>
         public string Package
         {
-            get { return this.package; }
-            set { this.package = value; }
+            get
+            {
+                if (!this.isLegacy)
+                {
+                    throw new ArgumentException("Package property does not exist");
+                }
+
+                return this.package;
+            }
+            set
+            {
+                if (!this.isLegacy)
+                {
+                    throw new ArgumentException("Package property does not exist");
+                }
+
+                this.package = value;
+            }
         }
 
         /// <summary>
         /// Gets or sets a value indicating whether the service should use verbose logging.
+        /// The property only exists in legacy mode.
         /// </summary>
         public bool UseVerboseLogging
         {
-            get { return this.useVerboseLogging; }
-            set { this.useVerboseLogging = value; }
+            get
+            {
+                if (!this.isLegacy)
+                {
+                    throw new ArgumentException("UseVerboseLogging property does not exist");
+                }
+
+                return this.useVerboseLogging;
+            }
+            set
+            {
+                if (!this.isLegacy)
+                {
+                    throw new ArgumentException("UseVerboseLogging property does not exist");
+                }
+
+                this.useVerboseLogging = value;
+            }
         }
 
         /// <summary>
         /// Gets or sets a value indicating whether the <see cref="EdgeDriverService"/> instance
         /// should use the a protocol dialect compliant with the W3C WebDriver Specification.
+        /// The property only exists in legacy mode.
         /// </summary>
         /// <remarks>
         /// Setting this property to a non-<see langword="null"/> value for driver
@@ -87,8 +144,24 @@ namespace OpenQA.Selenium.Edge
         /// </remarks>
         public bool? UseSpecCompliantProtocol
         {
-            get { return this.useSpecCompliantProtocol; }
-            set { this.useSpecCompliantProtocol = value; }
+            get
+            {
+                if (!this.isLegacy)
+                {
+                    throw new ArgumentException("UseVerboseLogging property does not exist");
+                }
+
+                return this.useSpecCompliantProtocol;
+            }
+            set
+            {
+                if (!this.isLegacy)
+                {
+                    throw new ArgumentException("UseVerboseLogging property does not exist");
+                }
+
+                this.useSpecCompliantProtocol = value;
+            }
         }
 
         /// <summary>
@@ -99,7 +172,7 @@ namespace OpenQA.Selenium.Edge
         {
             get
             {
-                if (this.useSpecCompliantProtocol.HasValue && !this.useSpecCompliantProtocol.Value)
+                if (!this.isLegacy || (this.useSpecCompliantProtocol.HasValue && !this.useSpecCompliantProtocol.Value))
                 {
                     return base.HasShutdown;
                 }
@@ -119,7 +192,7 @@ namespace OpenQA.Selenium.Edge
             // gets us to the termination point much faster.
             get
             {
-                if (this.useSpecCompliantProtocol.HasValue && !this.useSpecCompliantProtocol.Value)
+                if (!this.isLegacy || (this.useSpecCompliantProtocol.HasValue && !this.useSpecCompliantProtocol.Value))
                 {
                     return base.TerminationTimeout;
                 }
@@ -135,6 +208,9 @@ namespace OpenQA.Selenium.Edge
         {
             get
             {
+                if (!this.isLegacy)
+                    return base.CommandLineArguments;
+
                 StringBuilder argsBuilder = new StringBuilder(base.CommandLineArguments);
                 if (!string.IsNullOrEmpty(this.host))
                 {
@@ -175,11 +251,14 @@ namespace OpenQA.Selenium.Edge
         /// <summary>
         /// Creates a default instance of the EdgeDriverService.
         /// </summary>
+        /// <param name="isLegacy">Wheter to use legacy mode. Default is to true.</param>
         /// <returns>A EdgeDriverService that implements default settings.</returns>
-        public static EdgeDriverService CreateDefaultService()
+        public static EdgeDriverService CreateDefaultService(bool isLegacy = true)
         {
-            string serviceDirectory = DriverService.FindDriverServiceExecutable(MicrosoftWebDriverServiceFileName, MicrosoftWebDriverDownloadUrl);
-            EdgeDriverService service = CreateDefaultService(serviceDirectory);
+            string serviceDirectory = DriverService.FindDriverServiceExecutable(
+                isLegacy ? MicrosoftWebDriverServiceFileName : ChromiumDriverServiceFileName(MSEdgeDriverServiceFileName),
+                MicrosoftWebDriverDownloadUrl);
+            EdgeDriverService service = CreateDefaultService(serviceDirectory, isLegacy);
             return service;
         }
 
@@ -187,21 +266,25 @@ namespace OpenQA.Selenium.Edge
         /// Creates a default instance of the EdgeDriverService using a specified path to the EdgeDriver executable.
         /// </summary>
         /// <param name="driverPath">The directory containing the EdgeDriver executable.</param>
+        /// <param name="isLegacy">Wheter to use legacy mode. Default is to true.</param>
         /// <returns>A EdgeDriverService using a random port.</returns>
-        public static EdgeDriverService CreateDefaultService(string driverPath)
+        public static EdgeDriverService CreateDefaultService(string driverPath, bool isLegacy = true)
         {
-            return CreateDefaultService(driverPath, MicrosoftWebDriverServiceFileName);
+            return CreateDefaultService(
+                driverPath,
+                isLegacy ? MicrosoftWebDriverServiceFileName : ChromiumDriverServiceFileName(MSEdgeDriverServiceFileName),
+                isLegacy);
         }
 
-        /// <summary>
+        /// <summary> 
         /// Creates a default instance of the EdgeDriverService using a specified path to the EdgeDriver executable with the given name.
         /// </summary>
         /// <param name="driverPath">The directory containing the EdgeDriver executable.</param>
         /// <param name="driverExecutableFileName">The name of the EdgeDriver executable file.</param>
         /// <returns>A EdgeDriverService using a random port.</returns>
-        public static EdgeDriverService CreateDefaultService(string driverPath, string driverExecutableFileName)
+        public static EdgeDriverService CreateDefaultService(string driverPath, string driverExecutableFileName, bool isLegacy = true)
         {
-            return CreateDefaultService(driverPath, driverExecutableFileName, PortUtilities.FindFreePort());
+            return CreateDefaultService(driverPath, driverExecutableFileName, PortUtilities.FindFreePort(), isLegacy);
         }
 
         /// <summary>
@@ -210,10 +293,11 @@ namespace OpenQA.Selenium.Edge
         /// <param name="driverPath">The directory containing the EdgeDriver executable.</param>
         /// <param name="driverExecutableFileName">The name of the EdgeDriver executable file</param>
         /// <param name="port">The port number on which the driver will listen</param>
+        /// <param name="isLegacy">Wheter to use legacy mode. Default is to true.</param>
         /// <returns>A EdgeDriverService using the specified port.</returns>
-        public static EdgeDriverService CreateDefaultService(string driverPath, string driverExecutableFileName, int port)
+        public static EdgeDriverService CreateDefaultService(string driverPath, string driverExecutableFileName, int port, bool isLegacy = true)
         {
-            return new EdgeDriverService(driverPath, driverExecutableFileName, port);
+            return new EdgeDriverService(driverPath, driverExecutableFileName, port, isLegacy);
         }
     }
 }
