+++ dotnet/test/common/ClickTest.cs
@@ -340,6 +340,7 @@ namespace OpenQA.Selenium
         [Test]
         [IgnoreBrowser(Browser.IE, "Element is properly seen as obscured.")]
         [IgnoreBrowser(Browser.Chrome, "Element is properly seen as obscured.")]
+        [IgnoreBrowser(Browser.ChromiumEdge, "Element is properly seen as obscured.")]
         [IgnoreBrowser(Browser.Edge, "Element is properly seen as obscured.")]
         [IgnoreBrowser(Browser.Firefox, "Element is properly seen as obscured.")]
         [IgnoreBrowser(Browser.Safari, "Element is properly seen as obscured.")]
