+++ dotnet/test/common/ContentEditableTest.cs
@@ -64,6 +64,7 @@ namespace OpenQA.Selenium
 
         [Test]
         [IgnoreBrowser(Browser.Chrome, "Driver prepends text in contentEditable areas")]
+        [IgnoreBrowser(Browser.ChromiumEdge, "Driver prepends text in contentEditable areas")]
         [IgnoreBrowser(Browser.Firefox, "Driver prepends text in contentEditable areas")]
         [IgnoreBrowser(Browser.Safari, "Driver prepends text to contentEditable areas")]
         public void ShouldBeAbleToTypeIntoContentEditableElementWithExistingValue()
@@ -93,6 +94,7 @@ namespace OpenQA.Selenium
 
         [Test]
         [IgnoreBrowser(Browser.Chrome, "Driver prepends text in contentEditable areas")]
+        [IgnoreBrowser(Browser.ChromiumEdge, "Driver prepends text in contentEditable areas")]
         [IgnoreBrowser(Browser.Firefox, "Driver prepends text in contentEditable areas")]
         [IgnoreBrowser(Browser.IE, "Prepends text")]
         [IgnoreBrowser(Browser.Safari, "Driver prepends text to contentEditable areas")]
@@ -111,6 +113,7 @@ namespace OpenQA.Selenium
 
         [Test]
         [IgnoreBrowser(Browser.Chrome, "Driver prepends text in contentEditable areas")]
+        [IgnoreBrowser(Browser.ChromiumEdge, "Driver prepends text in contentEditable areas")]
         [IgnoreBrowser(Browser.Firefox, "Browser does not automatically focus body element in frame")]
         [IgnoreBrowser(Browser.Safari, "Driver prepends text to contentEditable areas")]
         public void AppendsTextToEndOfContentEditableWithMultipleTextNodes()
