+++ dotnet/test/common/CookieImplementationTest.cs
@@ -181,6 +181,7 @@ namespace OpenQA.Selenium
 
         [Test]
         [IgnoreBrowser(Browser.Chrome, "Chrome does not retrieve cookies when in frame.")]
+        [IgnoreBrowser(Browser.ChromiumEdge, "Edge does not retrieve cookies when in frame.")]
         public void GetCookiesInAFrame()
         {
             driver.Url = EnvironmentManager.Instance.UrlBuilder.WhereIs("animals");
