+++ dotnet/test/common/ElementFindingTest.cs
@@ -360,6 +360,7 @@ namespace OpenQA.Selenium
         [IgnoreBrowser(Browser.Edge, "Class name is perfectly legal when using CSS selector, if properly escaped.")]
         [IgnoreBrowser(Browser.Safari, "Class name is perfectly legal when using CSS selector, if properly escaped.")]
         [IgnoreBrowser(Browser.Chrome, "Class name is perfectly legal when using CSS selector, if properly escaped.")]
+        [IgnoreBrowser(Browser.ChromiumEdge, "Class name is perfectly legal when using CSS selector, if properly escaped.")]
         [IgnoreBrowser(Browser.Opera, "Throws WebDriverException")]
         public void FindingMultipleElementsByInvalidClassNameShouldThrow()
         {
@@ -445,6 +446,7 @@ namespace OpenQA.Selenium
         [Test]
         [IgnoreBrowser(Browser.IE, "Driver does not support finding elements on XML documents.")]
         [IgnoreBrowser(Browser.Chrome, "Driver does not support finding elements on XML documents.")]
+        [IgnoreBrowser(Browser.ChromiumEdge, "Driver does not support finding elements on XML documents.")]
         [IgnoreBrowser(Browser.Edge, "Driver does not support finding elements on XML documents.")]
         [IgnoreBrowser(Browser.Safari, "Not yet implemented")]
         public void ShouldBeAbleToFindElementByXPathInXmlDocument()
