+++ dotnet/test/common/MiscTest.cs
@@ -55,6 +55,7 @@ namespace OpenQA.Selenium
 
         [Test]
         [IgnoreBrowser(Browser.Chrome, "returns XML content formatted for display as HTML document")]
+        [IgnoreBrowser(Browser.ChromiumEdge, "returns XML content formatted for display as HTML document")]
         [IgnoreBrowser(Browser.Safari, "returns XML content formatted for display as HTML document")]
         [IgnoreBrowser(Browser.IE, "returns XML content formatted for display as HTML document")]
         [IgnoreBrowser(Browser.Edge, "returns XML content formatted for display as HTML document")]
