+++ dotnet/test/common/WindowSwitchingTest.cs
@@ -426,6 +426,7 @@ namespace OpenQA.Selenium
 
         [Test]
         [IgnoreBrowser(Browser.Chrome, "Driver does not yet support new window command")]
+        [IgnoreBrowser(Browser.ChromiumEdge, "Driver does not yet support new window command")]
         [IgnoreBrowser(Browser.Edge, "Driver does not yet support new window command")]
         public void ShouldBeAbleToCreateANewWindow()
         {
